Validator와 LadderGame구조체 각각의 테스트 함수를 분리한다
